+++ src/test/dummyData.ts
@@ -71,7 +71,11 @@ export const testPlants: Plant[] = [
     { id: 'Two', title: 'Test plant 2', slug: 'yet-another-slug' },
 ];
 
-export const dummyPermissions: string[] = ['COMMPKG/READ'];
+export const dummyPermissions: string[] = [
+    'COMMPKG/READ',
+    'CPCL/READ',
+    'PUNCHLISTITEM/READ',
+];
 
 export const testProjects: Project[] = [
     { id: 1, title: 'Test project 1', description: 'this-is-a-description' },
